requirements.txt
PyYAML
pytest
xxhash
Pillow